poker_cv_agent/requirements.txt
numpy==2.2.2
opencv-python==4.10.0.84
mss==10.0.0
pyautogui==0.9.54
pytesseract==0.3.13
PyYAML==6.0.2
treys==0.1.8
